intro: Module split, renames and documentation
- Split code into smaller modules
- Rename some of the modules to reflect better what they do
- Document things
