+++ introspection-engine/connectors/sql-introspection-connector/src/datamodel_calculator/context.rs
@@ -1,11 +1,10 @@
 use crate::{
-    introspection::introspect,
     introspection_helpers::{is_new_migration_table, is_old_migration_table, is_prisma_join_table, is_relay_table},
     introspection_map::RelationName,
     pair::{EnumPair, ModelPair, Pair, RelationFieldDirection},
-    warnings, EnumVariantName, IntrospectedName, ModelName, SqlFamilyTrait, SqlIntrospectionResult,
+    warnings, EnumVariantName, IntrospectedName, ModelName,
 };
-use introspection_connector::{IntrospectionContext, IntrospectionResult, Version, Warning};
+use introspection_connector::{Version, Warning};
 use psl::{
     builtin_connectors::*,
     datamodel_connector::Connector,
@@ -16,108 +15,14 @@ use quaint::prelude::SqlFamily;
 use sql_schema_describer as sql;
 use std::borrow::Cow;
 
-#[derive(Debug, Default)]
-pub(crate) struct Warnings {
-    pub(crate) warnings: Vec<Warning>,
-    pub(crate) prisma_1_uuid_defaults: Vec<warnings::ModelAndField>,
-    pub(crate) prisma_1_cuid_defaults: Vec<warnings::ModelAndField>,
-    pub(crate) fields_with_empty_names: Vec<warnings::ModelAndField>,
-    pub(crate) remapped_fields: Vec<warnings::ModelAndField>,
-    pub(crate) enum_values_with_empty_names: Vec<warnings::EnumAndValue>,
-    pub(crate) models_without_columns: Vec<warnings::Model>,
-    pub(crate) models_without_identifiers: Vec<warnings::Model>,
-    pub(crate) reintrospected_id_names: Vec<warnings::Model>,
-    pub(crate) unsupported_types: Vec<warnings::ModelAndFieldAndType>,
-    pub(crate) remapped_models: Vec<warnings::Model>,
-    pub(crate) reintrospected_relations: Vec<warnings::Model>,
+pub(crate) struct OutputContext<'a> {
+    pub(crate) rendered_schema: datamodel_renderer::Datamodel<'a>,
+    pub(crate) warnings: warnings::Warnings,
 }
 
-impl Warnings {
-    pub(crate) fn new() -> Self {
-        Self {
-            warnings: Vec::new(),
-            ..Default::default()
-        }
-    }
-
-    pub(crate) fn push(&mut self, warning: Warning) {
-        self.warnings.push(warning);
-    }
-
-    pub(crate) fn finalize(&mut self) -> Vec<Warning> {
-        fn maybe_warn<T>(elems: &[T], warning: impl Fn(&[T]) -> Warning, warnings: &mut Vec<Warning>) {
-            if !elems.is_empty() {
-                warnings.push(warning(elems))
-            }
-        }
-
-        maybe_warn(
-            &self.models_without_identifiers,
-            warnings::warning_models_without_identifier,
-            &mut self.warnings,
-        );
-
-        maybe_warn(
-            &self.unsupported_types,
-            warnings::warning_unsupported_types,
-            &mut self.warnings,
-        );
-
-        maybe_warn(
-            &self.remapped_models,
-            warnings::warning_enriched_with_map_on_model,
-            &mut self.warnings,
-        );
-
-        maybe_warn(
-            &self.remapped_fields,
-            warnings::warning_enriched_with_map_on_field,
-            &mut self.warnings,
-        );
-
-        maybe_warn(
-            &self.models_without_columns,
-            warnings::warning_models_without_columns,
-            &mut self.warnings,
-        );
-
-        maybe_warn(
-            &self.reintrospected_id_names,
-            warnings::warning_enriched_with_custom_primary_key_names,
-            &mut self.warnings,
-        );
-
-        maybe_warn(
-            &self.prisma_1_uuid_defaults,
-            warnings::warning_default_uuid_warning,
-            &mut self.warnings,
-        );
-
-        maybe_warn(
-            &self.prisma_1_cuid_defaults,
-            warnings::warning_default_cuid_warning,
-            &mut self.warnings,
-        );
-
-        maybe_warn(
-            &self.enum_values_with_empty_names,
-            warnings::warning_enum_values_with_empty_names,
-            &mut self.warnings,
-        );
-
-        maybe_warn(
-            &self.fields_with_empty_names,
-            warnings::warning_fields_with_empty_names,
-            &mut self.warnings,
-        );
-
-        maybe_warn(
-            &self.reintrospected_relations,
-            warnings::warning_relations_added_from_the_previous_data_model,
-            &mut self.warnings,
-        );
-
-        std::mem::take(&mut self.warnings)
+impl<'a> OutputContext<'a> {
+    pub(crate) fn finalize_warnings(&mut self) -> Vec<Warning> {
+        self.warnings.finalize()
     }
 }
 
@@ -130,17 +35,6 @@ pub(crate) struct InputContext<'a> {
     pub(crate) version: Version,
     pub(crate) previous_schema: &'a psl::ValidatedSchema,
     pub(crate) introspection_map: &'a crate::introspection_map::IntrospectionMap<'a>,
-}
-
-pub(crate) struct OutputContext<'a> {
-    pub(crate) rendered_schema: datamodel_renderer::Datamodel<'a>,
-    pub(crate) warnings: Warnings,
-}
-
-impl<'a> OutputContext<'a> {
-    pub(crate) fn finalize_warnings(&mut self) -> Vec<Warning> {
-        self.warnings.finalize()
-    }
 }
 
 impl<'a> InputContext<'a> {
@@ -392,49 +286,4 @@ impl<'a> InputContext<'a> {
                 (*direction, next)
             })
     }
-}
-
-/// Calculate a data model from a database schema.
-pub fn calculate_datamodel(
-    schema: &sql::SqlSchema,
-    ctx: &IntrospectionContext,
-) -> SqlIntrospectionResult<IntrospectionResult> {
-    let introspection_map = Default::default();
-
-    let mut input = InputContext {
-        version: Version::NonPrisma,
-        config: ctx.configuration(),
-        render_config: ctx.render_config,
-        schema,
-        sql_family: ctx.sql_family(),
-        previous_schema: ctx.previous_schema(),
-        introspection_map: &introspection_map,
-    };
-
-    let introspection_map = crate::introspection_map::IntrospectionMap::new(input);
-    input.introspection_map = &introspection_map;
-
-    let mut output = OutputContext {
-        rendered_schema: datamodel_renderer::Datamodel::default(),
-        warnings: Warnings::new(),
-    };
-
-    input.version = crate::version_checker::check_prisma_version(&input);
-
-    let (schema_string, is_empty) = introspect(input, &mut output)?;
-    let warnings = output.finalize_warnings();
-
-    // Warning codes 5 and 6 are for Prisma 1 default reintrospection.
-    let version = if warnings.iter().any(|w| ![5, 6].contains(&w.code)) {
-        Version::NonPrisma
-    } else {
-        input.version
-    };
-
-    Ok(IntrospectionResult {
-        data_model: schema_string,
-        is_empty,
-        version,
-        warnings,
-    })
 }
