+++ introspection-engine/connectors/sql-introspection-connector/src/pair/relation_field.rs
@@ -1,5 +1,5 @@
 use crate::{
-    calculate_datamodel::InputContext,
+    datamodel_calculator::InputContext,
     pair::{ModelPair, Pair},
 };
 use psl::{
